+++ src/main/kotlin/crimera/patches/twitter/premium/unlockdownloads/fingerprints/DownloadPatchFingerprint.kt
@@ -1,12 +1,17 @@
 package crimera.patches.twitter.premium.unlockdownloads.fingerprints
 
 import app.revanced.patcher.fingerprint.MethodFingerprint
+import com.android.tools.smali.dexlib2.Opcode
 
-object DownloadPatchFingerprint : MethodFingerprint(
-    returnType = "V",
+internal object DownloadPatchFingerprint : MethodFingerprint(
+    opcodes = listOf(
+        Opcode.IF_EQ,
+        Opcode.SGET_OBJECT,
+        Opcode.GOTO_16,
+        Opcode.NEW_INSTANCE,
+    ),
     strings = listOf(
+        "mediaEntity",
         "media_options_sheet",
-        "resources.getString(R.string.post_video)",
-        "resources.getString(R.string.post_photo)"
     )
 )